# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cart-abandonment:
    post:
      operationId: saveCartAbandonment
      tags: [analytics]
      summary: Save an anonymous or signed-in cart snapshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveCartAbandonmentInput"
      responses:
        "200":
          description: Saved cart snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AbandonedCart"
        "400": { description: Invalid cart snapshot }
  /cart-abandonment/{cartId}/recover:
    post:
      operationId: recoverCartAbandonment
      tags: [analytics]
      summary: Mark a cart as recovered after order creation
      parameters:
        - name: cartId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Recovery result
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { type: boolean }
        "400": { description: Invalid cart id }
...
  /admin/abandoned-carts:
    get:
      operationId: listAdminAbandonedCarts
      tags: [admin, analytics]
      summary: List carts inactive for at least one hour
      responses:
        "200":
          description: Abandoned cart list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AbandonedCart"
        "401": { description: Admin sign-in required }
...
components:
...
  schemas:
...
    AbandonedCartItem:
      type: object
      required: [productId, productName, duration, quantity, unitPrice]
      properties:
        productId: { type: number }
        productName: { type: string }
        duration: { type: string }
        quantity: { type: number }
        unitPrice: { type: number }
    AbandonedCart:
      type: object
      required: [id, cartId, customerId, customerName, customerEmail, customerPhone, currency, subtotal, itemCount, items, status, lastSeenAt, createdAt, recoveredAt]
      properties:
        id: { type: number }
        cartId: { type: string }
        customerId: { type: ["string", "null"] }
        customerName: { type: ["string", "null"] }
        customerEmail: { type: ["string", "null"] }
        customerPhone: { type: ["string", "null"] }
        currency:
          type: string
          enum: [EGP, USD]
        subtotal: { type: number }
        itemCount: { type: number }
        items:
          type: array
          items:
            $ref: "#/components/schemas/AbandonedCartItem"
        status:
          type: string
          enum: [open, recovered, expired]
        lastSeenAt: { type: string }
        createdAt: { type: string }
        recoveredAt: { type: ["string", "null"] }
    SaveCartAbandonmentInput:
      type: object
      required: [cartId, currency, subtotal, itemCount, items]
      properties:
        cartId: { type: string }
        currency:
          type: string
          enum: [EGP, USD]
        subtotal: { type: number, minimum: 0 }
        itemCount: { type: number, minimum: 1 }
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/AbandonedCartItem"
...