# lib/api-zod/src/generated/types/abandonedCart.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AbandonedCartCurrency } from './abandonedCartCurrency';
import type { AbandonedCartItem } from './abandonedCartItem';
import type { AbandonedCartStatus } from './abandonedCartStatus';

export interface AbandonedCart {
  id: number;
  cartId: string;
  /** @nullable */
  customerId: string | null;
  /** @nullable */
  customerName: string | null;
  /** @nullable */
  customerEmail: string | null;
  /** @nullable */
  customerPhone: string | null;
  currency: AbandonedCartCurrency;
  subtotal: number;
  itemCount: number;
  items: AbandonedCartItem[];
  status: AbandonedCartStatus;
  lastSeenAt: string;
  createdAt: string;
  /** @nullable */
  recoveredAt: string | null;
}