# lib/api-zod/src/generated/types/saveCartAbandonmentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AbandonedCartItem } from './abandonedCartItem';
import type { SaveCartAbandonmentInputCurrency } from './saveCartAbandonmentInputCurrency';

export interface SaveCartAbandonmentInput {
  cartId: string;
  currency: SaveCartAbandonmentInputCurrency;
  /** @minimum 0 */
  subtotal: number;
  /** @minimum 1 */
  itemCount: number;
  /** @minItems 1 */
  items: AbandonedCartItem[];
}